allow multiple simulations to be ran at the same time
