+++ bsb/simulation/adapter.py
@@ -1,5 +1,11 @@
 import abc
+import itertools
+import types
 import typing
+from contextlib import ExitStack
+from time import time
+
+import numpy as np
 
 from .results import SimulationResult
 
@@ -7,6 +13,33 @@ if typing.TYPE_CHECKING:
     from ..storage import PlacementSet
     from .cell import CellModel
     from .simulation import Simulation
+
+
+class AdapterProgress:
+    def __init__(self, duration):
+        self._progdur = duration
+        self._progstart = self._last_progtic = time()
+        self._progtics = 0
+
+    def progress(self, step):
+        """
+        Report simulation progress.
+        """
+        now = time()
+        tic = now - self._last_progtic
+        self._progtics += 1
+        el = now - self._progstart
+        progress = types.SimpleNamespace(
+            progression=step, duration=self._progdur, time=time(), tick=tic, elapsed=el
+        )
+        self._last_progtic = now
+        return progress
+
+    def step_progress(self, duration, step=1):
+        steps = itertools.chain(np.arange(0, duration), (duration,))
+        a, b = itertools.tee(steps)
+        next(b, None)
+        yield from zip(a, b)
 
 
 class SimulationData:
@@ -25,18 +58,29 @@ class SimulationData:
 
 class SimulatorAdapter(abc.ABC):
     def __init__(self):
+        self._progress_listeners = []
         self.simdata: dict["Simulation", "SimulationData"] = dict()
 
-    def simulate(self, simulation):
+    def simulate(self, *simulations, post_prepare=None, comm=None):
         """
-        Simulate the given simulation.
+        Simulate the given simulations.
         """
-        with simulation.scaffold.storage.read_only():
-            data = self.prepare(simulation)
-            for hook in simulation.post_prepare:
-                hook(self, simulation, data)
-            result = self.run(simulation)
-            return self.collect(simulation, data, result)
+        with ExitStack() as context:
+            for simulation in simulations:
+                context.enter_context(simulation.scaffold.storage.read_only())
+            alldata = []
+            for simulation in simulations:
+                data = self.prepare(simulation)
+                alldata.append(data)
+                for hook in simulation.post_prepare:
+                    hook(self, simulation, data)
+            if post_prepare:
+                post_prepare(self, simulations, alldata)
+            results = self.run(*simulations)
+            return [
+                self.collect(simulation, data, result)
+                for simulation, result in zip(simulations, results)
+            ]
 
     @abc.abstractmethod
     def prepare(self, simulation, comm=None):
@@ -51,15 +95,18 @@ class SimulatorAdapter(abc.ABC):
         pass
 
     @abc.abstractmethod
-    def run(self, simulation):
+    def run(self, *simulations, comm=None):
         """
         Fire up the prepared adapter.
         """
         pass
 
-    def collect(self, simulation, simdata, simresult):
+    def collect(self, simulation, simdata, simresult, comm=None):
         """
         Collect the output of a simulation that completed
         """
         simresult.flush()
         return simresult
+
+    def add_progress_listener(self, listener):
+        self._progress_listeners.append(listener)
